plots.jl

plots.jl

This unit implements utilities for plotting data produced by Fourier Analysis

RecipesBase.plot — Function.

(1)
plot( h      :: Taper,
    showkind :: Bool = false;
   args...)

(2)
plot(S      :: Spectra;
    maxf    :: Union{Real, Int} = 0,
    space   :: Int              = 4,
    ylabel  :: String = "Power (\muV²)",
   args...)

(1) Plot the data of a Taper object, that is, the tapering window(s) hold in its .y field. This is a single tapering window for all kind of tapering windows, but for Slepian's discrete prolate spheroid sequences, which are multi-tapering.

If optional keyword argument showkind is true, the output of the taperinfo function is printed as title of the plot. By defaut showkind is false.

(2) Produce the line plot of the data in a Spectra object, that is, plot all its spectra with separate lines.

Optional keyword argument maxf limits the plot on the right to the given frequency (in Hz).

Optional keyword argument space (in Hz) determines the spacing between frequency ticks and labels (x-axis).

Both method (1) and (2) allow to pass any other optional kyword argument of the standard Julia plot function.

See also: bar

Examples:

using FourierAnalysis, Plots
sr, t, f, a = 128, 128, 16, 0.5
# create a sinusoidal at 16Hz superimposed to white noise
v=sinusoidal(a, f, sr, t*16, 0) + randn(t*16)
# create a data matrix
X=broadcast(+, v, randn(t*16, 3)) * randn(3, 3)
# compute spectra using hamming tapering window
H=taper(hamming, t)
S=spectra(X, sr, t; tapering=H)

# plot tapering window
plot(H; title="my own title") # title is a Plot.jl argument
plot(H; showkind=true)

# plot spectra
plot(S)
plot(S; maxf=32)
plot(S; maxf=32, space=2)

# compute and plot amplitude spectra
S=spectra(X, sr, t; tapering=H, func=sqrt)
plot(S; maxf=32, space=2, ylabel="Amplitude (\muV)")

# plot Slepian's dpss
plot(slepians(sr, t*4, 2); showkind=true)
plot(slepians(sr, t*16, 1); title="my own title")

source

Plots.bar — Function.

 bar(S      :: Spectra;
     maxf    :: Union{Real, Int} = 0,
     space   :: Int              = 4,
     ylabel  :: String = "Power (\muV²)",
    args...)

Produce the bar plot of the data in a Spectra object, that is, plot all its spectra with separate lines. Use the same syntax as the plot fuction, which produces a line plot instead. Bar plots for spectra can be used insted of the line plots only for univariate spectra.

See: plot

source

Plots.heatmap — Function.

function heatmap(Y    :: TFAnalyticSignal,
                 func :: Function;
            c :: Symbol = :default,
            args...)

Generate heatmaps from a [TFAnalyticSignal] object. Since the analytic signal (AS) is complex, a function func must be specified and must transform the AS data matrix into a real matrix. Typical functions are:

real: heatmap of the real part of the AS

imag: heatmap of the imaginary part of the AS

amplitude: heatmap of the AS amplitude

phase: heatmap of the AS phase

Note:

Appropriate colors for plotting analytic signal and phase are :pu_or, :bluesreds. Appropriate colors for plotting amplitude are :amp, :fire, :dimgray, :gwv.

Examples:

using FourierAnalysis, Plots
sr, t, f, a = 128, 128, 8, 2
# create a sinusoidal at 8Hz
wave=sinusoidal(a, f, sr, t*4, 0)
# create a modulating sinusoidal at 0.25Hz
modu=sinusoidal(a, 0.25, sr, t*4, 0)
# create modulated wave and superimpose white noise
v=wave .* modu + randn(t*4)
plot(v)

# compute analytic signal (AS) of vector v
Y=TFanalyticsignal(v, sr, t; fmax=32)

# plot the real part of the AS
heatmap(Y, real)

# ...the imaginary part of the AS
heatmap(Y, imag)

# ...the amplitude of the AS
heatmap(Y, amplitude)

# ...the amplitude of the AS smoothed in the freq. dim.
heatmap(smooth(hannSmoother, noSmoother, Y), amplitude)

# ...the amplitude of the AS smoothed in freq. and time
heatmap(smooth(hannSmoother, hannSmoother, Y), amplitude)

# ...the phase of the AS and use custom colors
heatmap(Y, phase; c=:pu_or)

source

function heatmap(Y :: TFAmplitude;
        c :: Symbol = :default,
        args...) =

Generate heatmaps of the data in a TFAmplitude object.

Note: appropriate colors for plotting amplitude are :amp, :fire, :dimgray, :gwv.

Examples:

using FourierAnalysis, Plots
sr, t, f, a = 128, 128, 8, 2
# create a sinusoidal at 8Hz
wave=sinusoidal(a, f, sr, t*4, 0)
# create a modulating sinusoidal at 0.25Hz
modu=sinusoidal(a, 0.25, sr, t*4, 0)
# create modulated wave and superimpose white noise
v=wave .* modu + randn(t*4)
plot(v)

# compute amplitude of vector v
A=TFamplitude(v, sr, t; fmax=32)

heatmap(A)
# or, quicly: heatmap(TFamplitude(v, sr, t; fmax=32))

heatmap(A; c=:fire)

source

function heatmap(Y :: TFPhase;
              c :: Symbol = :default,
              args...)

Generate heatmaps of the data in a TFPhase object. Default colors for the heatmap are c=:amp if the phase object is unwrapped, c=:bluesreds otherwise.

Note:

Appropriate colors for plotting phase: :pu_or, :bluesreds Appropriate colors for plotting unwrapper phase are :amp, :fire, :dimgray, :gwv.

See also: isUnwrapped

Examples:

using FourierAnalysis, Plots
sr, t, f, a = 128, 128, 8, 2
# create a sinusoidal at 8Hz
wave=sinusoidal(a, f, sr, t*4, 0)
# create a modulating sinusoidal at 0.25Hz
modu=sinusoidal(a, 0.25, sr, t*4, 0)
# create modulated wave and superimpose white noise
v=wave .* modu + randn(t*4)
plot(v)

# heatmap of the phase of vector v
heatmap(TFphase(v, sr, t; fmax=32))

# heatmap of the unwrapped phase of vector v
heatmap(unwrapPhase(TFphase(v, sr, t; fmax=32)); c=:fire)

source